requirements.txt
click==8.0.3
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
flask-nav==0.6
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysqlclient==2.0.3
PyYAML==6.0
visitor==0.1.3
Werkzeug==2.0.2
